Resolve review comments - move the util function to metrics_utils.ts.
